Almost done
Go to the CovidReader class and go all the way to the bottom to where the print() method is. This is the method that is going to print out our code to match the output files. In the print method, everything is all done I just need you to replace the two comments with what they say. A races variable is already created i just need you to sort that variable